+++ src/php/Console/Infrastructure/Command/RunCommands.php
@@ -7,6 +7,7 @@ namespace Phel\Console\Infrastructure\Command;
 use Phel\Console\Domain\ConsoleCommandProviderInterface;
 use Phel\Run\Infrastructure\Command\AgentInstallCommand;
 use Phel\Run\Infrastructure\Command\CompileCommand;
+use Phel\Run\Infrastructure\Command\ConfigCommand;
 use Phel\Run\Infrastructure\Command\DoctorCommand;
 use Phel\Run\Infrastructure\Command\EvalCommand;
 use Phel\Run\Infrastructure\Command\InitCommand;
@@ -31,6 +32,7 @@ final class RunCommands implements ConsoleCommandProviderInterface
             new TestCommand(),
             new TestWorkerCommand(),
             new DoctorCommand(),
+            new ConfigCommand(),
         ];
     }
 }